Use AES CTR mode to encrypt blob on disk
DECRYPT-BLOB-AS-OCTETS will (potentially expensively) return the
unecrypted contents of a given blob as octets.  The method of
construction of the result is sub-optimially one byte at a time.

FIXME: Test of encrypt/decrypt of "/etc/passwd" into blob shows we are
somehow reading and/or writing an extra byte.

Add prototype to blob to enable JSON deserialization.

MAKE-NEW-DIRECTORY will (hopefully) portably return a new directory
under the value of the special *BLOBS-DIRECTORY*.

Bypassed CHUTE:GET-CIPHER in order to deal directly with ciphers
(needs rethinking).  GET-KEY gets a class encapsulating a keyed cipher
state.