Abstract ZFS and BTRFS implementations
In using the CL:DEFPACKAGE form as initial system entity sorting
attempt, the following packages are to be used for symbols.  All
symbols used between packages shall be exported and constitute the API
supported by that version of the package.

The CHUTE/CONFIG package contains all necessary configuration
information.

All cryptographic functions are to be accessed via CHUTE/CRYPT.

Snapshoting filesytems are to be used from CHUTE/FS which in turn uses
the implementations in CHUTE/ZFS and CHUTE/BTRFS as necessary. Model
ZFS and BTRFS as implementations.

Abstract filesystem into multiple implementations (BTRFS/ZFS).  The
plan of attack is to provide searchable metadata caches via
RUN-PROGRAM invocation of external utilities such as `mtree`.

Anchor URI bootpoints as URI to be resolved in setting up
decentralized services.